minor changes to curvature testing.

+++ Devel/Curvature-test/compare-only-inf.sh
@@ -12,7 +12,9 @@ ndepth=$1
 dim=$2'D'
 
 cp gerris-3D.txt gerris-3D.tmp
-cp ../../Tests/Curvature/paris-$dim-$ndepth.tmp . 
+if [ -f ../../Tests/Curvature/paris-$dim-$ndepth.tmp ] ; then
+    cp ../../Tests/Curvature/paris-$dim-$ndepth.tmp .
+fi
 cp paris-$dim-$ndepth.tmp 1.tmp
 egrep -v nan 1.tmp > paris-$dim-$ndepth.tmp
 rm -f 1.tmp
@@ -38,7 +40,7 @@ set grid
 set xrange [1:110]
 set xlabel "Grid points per Diameter"
 set ylabel "Curvature error in $dim"
-plot 2/(x*x) t "2/(x*x)", "bsk-$dim.txt" u (2*\$1):2 t "Basilisk-$dim max max" w lp, "paris-$dim-$ndepth-nocent.tmp" u (2*\$1):3 t "ParisSim max max HF + HF fit" w lp
+plot 2/(x*x) t "2/(x*x)", "bsk-$dim-log.txt" u 1:4 t "Basilisk-$dim max max" w lp, "paris-$dim-$ndepth-nocent.tmp" u (2*\$1):3 t "ParisSim max max HF + HF fit" w lp, "paris-$dim-$ndepth-nocent.txt" u (2*\$1):3 t "ParisSim max max HF + HF fit old" w lp, "paris-$dim-5-nocent.tmp" u (2*\$1):3 t "ParisSim max max HF + HF fit depth=5" w lp
 EOF
 fi
